ui additions, peldr and kernel init

+++ nearldr/src/externheaders/efi/Uefi.h
@@ -20,10 +20,11 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #ifndef __PI_UEFI_H__
 #define __PI_UEFI_H__
 
+#include "ProcessorBind.h"
 #include "UefiBaseType.h"
 #include "UefiSpec.h"
 
-typedef int wchar;
+typedef UINT16 wchar;
 
 #endif
